scripts/python/requirements.txt
dbt-snowflake>=1.7.0,<2.0.0
snowflake-connector-python>=3.8.0,<4.0.0
pandas>=2.0.0,<3.0.0
markdown>=3.5.2,<4.0.0
faker>=25.0.0,<26.0.0
pre-commit>=3.6.0,<4.0.0
sqlfluff==3.0.3